+++ src/Label/LabelGenerator.php
@@ -23,4 +23,11 @@ interface LabelGenerator
      * @return string[]
      */
     public function generate(Sequence $sequence): array;
+
+    /**
+     * Returns a formatted label according to the generator.
+     *
+     * @param mixed $str a stringable label
+     */
+    public function format($str): string;
 }